refactor(k8s-ui): share TimelineSwimlanes + Prometheus charts with the hub
Move the two tab presentational components that lived only in radar/web into the
shared @skyhook-io/k8s-ui so Radar Hub can reuse them (the pattern ResourcesView/
IssuesView/logs/topology already follow). Behavior-preserving for radar OSS.

- TimelineSwimlanes (swimlane + scrollable event table + cross-highlight) moves
  to k8s-ui/components/timeline. Its three radar-only couplings become props:
  hasLimitedAccess (was a context) and onNavigatePath (was react-router); the
  self-package import becomes relative. radar/web's TimelineSwimlanes is now a
  thin host wrapper injecting those — TimelineView and OSS behavior unchanged.
- New k8s-ui/components/charts/PrometheusChartsView: the pure chart UI (category
  tabs, time range, connect CTA, loading/empty/error states) over the existing
  AreaChart/MetricsSummary/SeriesLegend primitives. radar/web's PrometheusCharts
  keeps its fetch hooks, state, exports, and four consumers, and delegates only
  its render to the shared view.

tsc green; no behavior change to radar OSS. Publish as k8s-ui-v1.8.0; Radar Hub
then bumps its dependency to consume both components fed by its own tunnel data.
