+++ packages/k8s-ui/src/components/charts/index.ts
@@ -1,6 +1,19 @@
 export { AreaChart } from './AreaChart'
 export { MetricsSummary } from './MetricsSummary'
 export { SeriesLegend } from './SeriesLegend'
+export {
+  PrometheusChartsView,
+  WORKLOAD_METRIC_CATEGORIES,
+  NODE_METRIC_CATEGORIES,
+  METRIC_TIME_RANGES,
+} from './PrometheusChartsView'
+export type {
+  PrometheusChartsViewProps,
+  PrometheusMetricCategory,
+  PrometheusTimeRange,
+  MetricCategoryDef,
+  PrometheusResourceMetricsResult,
+} from './PrometheusChartsView'
 export { SERIES_COLORS, seriesColor, seriesFill, computeShortLabels } from './colors'
 export { formatMetricValue, formatTimestamp } from './format'
 export { computeSaturation } from './saturation'
