+++ packages/k8s-ui/src/components/timeline/index.ts
@@ -1,3 +1,4 @@
 export * from './shared'
 export * from './DiffViewer'
 export * from './TimelineList'
+export * from './TimelineSwimlanes'
